requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
gunicorn==22.0.0
bcrypt==4.2.0
flask-compress
requests==2.32.3
psycopg2-binary==2.9.10
openpyxl==3.1.5
fpdf2==2.7.6
matplotlib==3.8.2
pymysql==1.1.0
cryptography==42.0.5
httpx>=0.27.0
python-dotenv>=1.0.0
docling==2.64.1
fastapi==0.115.0
uvicorn[standard]==0.30.1
pgvector==0.2.4
prometheus-fastapi-instrumentator==6.0.0
langgraph==1.0.5